Add files via upload
This is the final practice with all the files needed for it to work. First my report, to show that I've worked out all the examples and things needed, then the server the main program of the final practice, the index there it is all the html code so the web page can be properly showed. 
the client json to try all the json code made. The error page in case the user introduces something that does not match with the web page abilities and finally the Seq program as we did some labs before.